minor file structure changes
