+++ automation/create_lab_repo/create_gitignore.py
@@ -1,6 +1,4 @@
-import os 
-import shutil
-import platform
+import os
 
 def create_gitignore(directory='.'):
     """
