+++ automation/main.py
@@ -1,11 +1,5 @@
-from automation.create_lab_repo.create_local_repo import init_local_repo
-from automation.create_lab_repo.create_gitignore import create_gitignore
-from automation.create_lab_repo.create_github_repo import create_github_repo
-from automation.create_lab_repo.user_prompts import user_prompts
-from automation.create_lab_repo.create_pip_install import create_pip_install
-from automation.create_lab_repo.create_readme import create_readme
+from automation.create_lab_repo.create_lab_repo import create_lab_repo
 from automation.readings_scraper.readings_scraper import readings_scraper
-from automation.create_lab_repo.check_gh_stuff import check_gh_user, check_gh_repo_exists
 from rich.console import Console
 from rich.prompt import Prompt
 
@@ -25,19 +19,6 @@ def main():
             readings_scraper()
         else:
             break
-    
-
-def create_lab_repo():
-    """
-    Function to create a lab repository with necessary files and configurations.
-    """
-    # Function calls
-    directory, username, repo_name, pip_installs = user_prompts()
-    init_local_repo(directory, repo_name)
-    create_readme(directory)
-    create_pip_install(pip_installs, directory)
-    create_gitignore(directory)
-    create_github_repo(repo_name, username, directory)
 
 
 if __name__ == "__main__":
