+++ tests/test_all.py
@@ -2,7 +2,7 @@ import pytest
 import os
 import shutil
 from unittest.mock import patch
-from automation.create_lab_repo.create_local_repo import init_local_repo
+from automation.create_lab_repo.init_local_repo import init_local_repo
 from automation.create_lab_repo.create_gitignore import create_gitignore
 from automation.create_lab_repo.create_readme import create_readme
 from automation.create_lab_repo.check_gh_stuff import check_gh_user, check_gh_repo_exists
